requirements.txt
faker
google-cloud-pubsub
apache-beam[gcp]
google-cloud-bigquery
dbt-bigquery
python-dotenv